Optimizations and bug fixes
