+++ vitrix/lib/items/aid_kit.py
@@ -12,7 +12,5 @@ class AidKit(Entity):
             scale=0.4,
         )
 
-        self.is_crate = False
-        self.is_aid_kit = True
         self.collider = MeshCollider(self)
         self.health_restore = random.randint(50, 80) # amount of health to give to player
